refactor(controller): extract flight queries to dedicated class, reducing code complexity
